Securizamos nuestros endpoints de API de usuarios

+++ users/api.py
@@ -4,10 +4,13 @@ from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from users.permissions import UserPermission
 from users.serializers import UserSerializer, UserListSerializer
 
 
 class UsersListAPIView(APIView):
+
+    permission_classes = [UserPermission]
 
     def get(self, request):
         users = User.objects.all()
@@ -24,19 +27,24 @@ class UsersListAPIView(APIView):
 
 class UserDetailAPIView(APIView):
 
+    permission_classes = [UserPermission]
+
     def get(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(user)
         return Response(serializer.data)
 
     def put(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(user, data=request.data)
         serializer.is_valid(raise_exception=True)  # devuelve una respuesta 400 BAD REQUEST
         serializer.save()
         return Response(serializer.data)
 
-    def delete(self, resquest, pk):
+    def delete(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
